Authorization

+++ web/app/Http/Controllers/JobController.php
@@ -3,6 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -14,14 +17,15 @@ class JobController extends Controller
             'jobs' => $jobs
         ]);
     }
+
     public function create()
     {
         return view('jobs.create');
     }
+
     public function show(Job $job)
     {
         return view('jobs.show', ['job' => $job]);
-
     }
 
     public function store()
@@ -30,40 +34,41 @@ class JobController extends Controller
             'title' => ['required', 'min:3'],
             'salary' => ['required']
         ]);
-    
+
         Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
-        ]);   
-        
-    return redirect('/jobs'); 
+        ]);
 
-     
+        return redirect('/jobs');
     }
+
     public function edit(Job $job)
     {
         return view('jobs.edit', ['job' => $job]);
     }
+
     public function update(Job $job)
-    { 
+    {
+        Gate::authorize('edit-job', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required']
         ]);
-    
-        // authorize (On hold...)
-    
+
         $job->update([
             'title' => request('title'),
             'salary' => request('salary'),
         ]);
-        return redirect('/jobs/' . $job->id);
-    
 
+        return redirect('/jobs/' . $job->id);
     }
+
     public function destroy(Job $job)
     {
+        Gate::authorize('edit-job', $job);
 
         $job->delete();
